backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.41
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.1.0
apscheduler==3.10.1
pillow==10.2.0
torch==2.1.0
transformers==4.36.2
numpy==1.24.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.11.3
pytest==7.4.0
httpx==0.24.1
tokenizers==0.15.0 